Funciones productos listo

+++ models/comands.go
@@ -7,7 +7,7 @@ type Product struct {
 	ID          uint    `gorm:"primaryKey" json:"id"`          // Identificador único del producto
 	Name        string  `gorm:"not null;unique" json:"name"`   // Nombre del producto
 	Description string  `gorm:"not null" json:"description"`   // Descripción del producto
-	Price       float64 `gorm:"not null" json:"price"`         // Precio del producto
+	Price       int `gorm:"not null" json:"price"`         // Precio del producto
 }
 
 // Representa un item dentro de una comanda (similar a un CartItem).
@@ -18,7 +18,7 @@ type OrderItem struct {
 	ProductID  uint    `gorm:"not null" json:"product_id"`    // Producto asociado
 	Product    Product `gorm:"foreignKey:ProductID" json:"product"` // Detalles del producto
 	Quantity   int     `gorm:"not null" json:"quantity"`      // Cantidad solicitada
-	TotalPrice float64 `gorm:"not null" json:"total_price"`   // Total del producto (Price * Quantity)
+	TotalPrice int `gorm:"not null" json:"total_price"`   // Total del producto (Price * Quantity)
 }
 
 // Representa una comanda en el restaurante.
@@ -28,7 +28,7 @@ type Order struct {
 	UserID       uint       `gorm:"not null" json:"user_id"`           // Usuario asignado a la comanda
 	Items        []OrderItem `gorm:"foreignKey:OrderID" json:"items"`         // Productos incluidos en la comanda
 	OrderDate    time.Time  `gorm:"not null" json:"order_date"`        // Fecha de creación del pedido
-	TotalAmount  float64    `gorm:"not null" json:"total_amount"`      // Total de la comanda
+	TotalAmount  int    `gorm:"not null" json:"total_amount"`      // Total de la comanda
 }
